Figure out how to lay out the redis stuff.
