+++ requirements.txt
@@ -26,6 +26,7 @@ pystache==0.5.4
 pytest==3.6.0
 python-dateutil==2.7.3
 pytz==2018.4
+redis==2.10.6
 requests==2.18.4
 requests_oauthlib==0.8.0
 seaborn==0.8.1
